refactor: 레스토랑 클래스 삭제
