--- src/domain/Restaurant.ts
@@ -1,11 +1,0 @@
-import { RestaurantType, CategoryType } from '../types';
-
-class Restaurant {
-  #restaurant;
-
-  constructor(restaurant: RestaurantType) {
-    this.#restaurant = restaurant;
-  }
-}
-
-export default Restaurant;
