+++ src/types/index.ts
@@ -1,4 +1,4 @@
-export { default as RestaurantType } from './restaurant';
+export { default as Restaurant } from './restaurant';
 export { default as CategoryType } from './category';
 export { default as SortType } from './sort';
 export { default as MatzipInterface } from './matzip';
